rename projects-copy to projects and adjust paths in files
