+++ controller/npm.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const changeCase = require('change-case');
 const greenlight = require('./greenlight');
-const projects = require('./projects-copy');
+const projects = require('./projects');
 const inquirer = require('inquirer');
 const exec = require('child_process').exec;
 const env = require('./env');
